Version 1.0.0
Added the current (extremely) beta version of the game. Base systems include:
Inventory System
Pick up Items
Interactables
It should be noted that these Systems are slightly independent of one another, and do not have full functionality with one another. Current build will allow the player to walk up to an item, be prompted to pick up with [e] and then be able to pick up. (NOTE: some items have not been set to destroy after pick up to test multiple items). There is also a door that can be interacted with, but a prompt will not be present.

I have also updated the gitignore to include the meta files.